+++ examples/ui5-ts-app/.gitignore
@@ -1,5 +1,4 @@
 # build results
-webapp
 dist
 
 # Logs
