+++ examples/ui5-ts-app/webapp/controller/App.controller.ts
@@ -1,4 +1,4 @@
-import BaseController from "./BaseController"
+import BaseController from "./BaseController.js"
 
 /**
  * @namespace test.Sample.tsapp.controller
