+++ examples/ui5-ts-app/webapp/controller/Main.controller.ts
@@ -1,4 +1,4 @@
-import BaseController from "./BaseController"
+import BaseController from "./BaseController.js"
 import oDataModel from "sap/ui/model/odata/v2/ODataModel"
 
 /**
