+++ examples/ui5-ts-app/webapp/i18n/i18n_en.properties
