+++ examples/ui5-ts-app/webapp/manifest.json
