+++ examples/ui5-ts-app/webapp/view/App.view.xml
