+++ examples/ui5-ts-app/webapp/view/Dialog.fragment.xml
