+++ examples/ui5-ts-app/webapp/view/Main.view.xml
