+++ examples/ui5-ts-app/webapp/view/Other.view.xml
